+++ docs/components/Nav/index.tsx
@@ -1,1 +1,2 @@
 export { DesktopNav } from './Desktop/DesktopNav';
+export { MobileNav } from './Mobile/MobileNav';