+++ docs/components/index.tsx
@@ -1,0 +1,29 @@
+export { AnchorLink } from './AnchorLink';
+export { APIheading } from './APIheading';
+export { FlexTable } from './FlexTable';
+export { List } from './List';
+export { Section } from './Section';
+export { SectionTitle } from './SectionTitle';
+export * from './Typography';
+export { Param, Params, ParamsDescription, ParamsTable, Syntax } from './Code';
+
+export { Content, Doc, Title, ArticleSubSectionTitle } from './Article';
+export type { ContentProps, DocProps, TitleProps } from './Article';
+
+export {
+  Brand,
+  Footer,
+  Header,
+  NavTop,
+  Tagline,
+  Container,
+  GlobalStyles,
+  HR,
+  HTMLhead,
+  Layout,
+  Spinner,
+  DottedSpinner
+} from './Global';
+
+export { Logo } from './Logo';
+export { DesktopNav, MobileNav } from './Nav';