+++ packages/flex/src/FlexCol.tsx
@@ -15,6 +15,9 @@ export const FlexCol = React.forwardRef((props: any, ref) => {
       flexBasis={columnSizing}
       pl={gutterWidth}
       pr={gutterWidth}
+      style={{
+        boxSizing: 'border-box'
+      }}
       ref={ref}
       {...props}
     >
@@ -34,8 +37,10 @@ export function setColumnSizing(
   columnSize: number | number[]
 ): string | string[] {
   return Array.isArray(columnSize) && columnSize.length > 1
-    ? columnSize.map(size => `${size * baseColumnWidth}%`)
-    : `${(columnSize as number) * baseColumnWidth}%`;
+    ? columnSize.map(
+        size => `${parseNumberToTwoDecimals(size * baseColumnWidth)}%`
+      )
+    : `${parseNumberToTwoDecimals((columnSize as number) * baseColumnWidth)}%`;
 }
 
 export function setColumnOffset(
@@ -59,6 +64,10 @@ export function setGutterWidth(
   return Array.isArray(gutterWidth) && gutterWidth.length > 1
     ? gutterWidth.map(width => `${width / 2}em`)
     : `${(gutterWidth as number) / 2}em`;
+}
+
+export function parseNumberToTwoDecimals(value: number) {
+  return `${parseFloat(String(value)).toFixed(2)}`;
 }
 
 export function isNumber(val: number): boolean {