+++ packages/media/src/mediaBoundaries.ts
@@ -58,7 +58,7 @@ export function createBoundaryString(
 }
 
 const OPERATOR_REGEX = /(?:[(>|<)=)]+)/g;
-const VALUE_REGEX = /((\d{1,}\/\d{1,})|(\d{1,}(r|em)|px|vh|vw))+/g;
+const VALUE_REGEX = /((\d{1,}\/\d{1,})|(\d{1,}(rem|em|px|vh|vw)))+/g;
 
 export const extractValues = extractByRegex.bind(null, VALUE_REGEX);
 export const extractOperators = extractByRegex.bind(null, OPERATOR_REGEX);